refactor(filesystem): Split web filesystem into smaller files
