+++ crates/filesystem/src/web/mod.rs
@@ -1,0 +1,408 @@
+// Copyright (C) 2023 Lily Lyons
+//
+// This file is part of Luminol.
+//
+// Luminol is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Luminol is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Luminol.  If not, see <http://www.gnu.org/licenses/>.
+
+mod events;
+mod util;
+pub use events::setup_main_thread_hooks;
+
+use super::FileSystem as FileSystemTrait;
+use super::{DirEntry, Error, Metadata, OpenFlags, Result};
+
+static FILESYSTEM_TX: once_cell::sync::OnceCell<flume::Sender<FileSystemCommand>> =
+    once_cell::sync::OnceCell::new();
+
+#[derive(Debug)]
+pub struct FileSystem {
+    key: usize,
+    name: String,
+    idb_key: Option<String>,
+}
+
+#[derive(Debug)]
+pub struct File {
+    key: usize,
+}
+
+#[derive(Debug)]
+pub struct FileSystemCommand(FileSystemCommandInner);
+
+#[derive(Debug)]
+enum FileSystemCommandInner {
+    Supported(oneshot::Sender<bool>),
+    DirEntryMetadata(
+        usize,
+        camino::Utf8PathBuf,
+        oneshot::Sender<Result<Metadata>>,
+    ),
+    DirPicker(oneshot::Sender<Option<(usize, String, Option<String>)>>),
+    DirFromIdb(String, oneshot::Sender<Option<(usize, String)>>),
+    DirSubdir(
+        usize,
+        camino::Utf8PathBuf,
+        oneshot::Sender<Result<(usize, String, Option<String>)>>,
+    ),
+    DirIdbDrop(String, oneshot::Sender<bool>),
+    DirOpenFile(
+        usize,
+        camino::Utf8PathBuf,
+        OpenFlags,
+        oneshot::Sender<Result<usize>>,
+    ),
+    DirEntryExists(usize, camino::Utf8PathBuf, oneshot::Sender<bool>),
+    DirCreateDir(usize, camino::Utf8PathBuf, oneshot::Sender<Result<()>>),
+    DirRemoveDir(usize, camino::Utf8PathBuf, oneshot::Sender<Result<()>>),
+    DirRemoveFile(usize, camino::Utf8PathBuf, oneshot::Sender<Result<()>>),
+    DirReadDir(
+        usize,
+        camino::Utf8PathBuf,
+        oneshot::Sender<Result<Vec<DirEntry>>>,
+    ),
+    DirDrop(usize, oneshot::Sender<bool>),
+    DirClone(usize, oneshot::Sender<usize>),
+    FileRead(usize, usize, oneshot::Sender<std::io::Result<Vec<u8>>>),
+    FileWrite(usize, Vec<u8>, oneshot::Sender<std::io::Result<()>>),
+    FileFlush(usize, oneshot::Sender<std::io::Result<()>>),
+    FileSeek(
+        usize,
+        std::io::SeekFrom,
+        oneshot::Sender<std::io::Result<u64>>,
+    ),
+    FileDrop(usize, oneshot::Sender<bool>),
+    FileSize(usize, oneshot::Sender<u64>),
+}
+
+fn filesystem_tx_or_die() -> &'static flume::Sender<FileSystemCommand> {
+    FILESYSTEM_TX.get().expect("FileSystem sender has not been initialized! Please call `FileSystem::initialize_sender` before calling this function.")
+}
+
+impl FileSystem {
+    /// Initializes the sender that we use to send filesystem commands to the main thread.
+    /// This must be called before performing any filesystem operations.
+    pub fn setup_filesystem_sender(filesystem_tx: flume::Sender<FileSystemCommand>) {
+        FILESYSTEM_TX
+            .set(filesystem_tx)
+            .expect("FileSystem sender cannot be initialized twice");
+    }
+
+    /// Returns whether or not the user's browser supports the JavaScript File System API.
+    pub fn filesystem_supported() -> bool {
+        let (oneshot_tx, oneshot_rx) = oneshot::channel();
+        filesystem_tx_or_die()
+            .send(FileSystemCommand(FileSystemCommandInner::Supported(
+                oneshot_tx,
+            )))
+            .unwrap();
+        oneshot_rx.recv().unwrap()
+    }
+
+    /// Attempts to prompt the user to choose a directory from their local machine using the
+    /// JavaScript File System API.
+    /// Then creates a `FileSystem` allowing read-write access to that directory if they chose one
+    /// successfully.
+    /// If the File System API is not supported, this always returns `None` without doing anything.
+    pub async fn from_folder_picker() -> Result<Self> {
+        if !Self::filesystem_supported() {
+            return Err(Error::Wasm32FilesystemNotSupported);
+        }
+        let (oneshot_tx, oneshot_rx) = oneshot::channel();
+        filesystem_tx_or_die()
+            .send(FileSystemCommand(FileSystemCommandInner::DirPicker(
+                oneshot_tx,
+            )))
+            .unwrap();
+        oneshot_rx
+            .await
+            .unwrap()
+            .map(|(key, name, idb_key)| FileSystem { key, name, idb_key })
+            .ok_or(Error::CancelledLoading)
+    }
+
+    /// Attempts to restore a previously created `FileSystem` using its `.idb_key()`.
+    pub async fn from_idb_key(idb_key: String) -> Result<Self> {
+        if !Self::filesystem_supported() {
+            return Err(Error::Wasm32FilesystemNotSupported);
+        }
+        let (oneshot_tx, oneshot_rx) = oneshot::channel();
+        filesystem_tx_or_die()
+            .send(FileSystemCommand(FileSystemCommandInner::DirFromIdb(
+                idb_key.clone(),
+                oneshot_tx,
+            )))
+            .unwrap();
+        oneshot_rx
+            .await
+            .unwrap()
+            .map(|(key, name)| FileSystem {
+                key,
+                name,
+                idb_key: Some(idb_key),
+            })
+            .ok_or(Error::MissingIDB)
+    }
+
+    /// Creates a new `FileSystem` from a subdirectory of this one.
+    pub fn subdir(&self, path: impl AsRef<camino::Utf8Path>) -> Result<Self> {
+        let (oneshot_tx, oneshot_rx) = oneshot::channel();
+        filesystem_tx_or_die()
+            .send(FileSystemCommand(FileSystemCommandInner::DirSubdir(
+                self.key,
+                path.as_ref().to_path_buf(),
+                oneshot_tx,
+            )))
+            .unwrap();
+        oneshot_rx
+            .recv()
+            .unwrap()
+            .map(|(key, name, idb_key)| FileSystem { key, name, idb_key })
+    }
+
+    /// Drops the directory with the given key from IndexedDB if it exists in there.
+    pub fn idb_drop(idb_key: String) -> bool {
+        let (oneshot_tx, oneshot_rx) = oneshot::channel();
+        filesystem_tx_or_die()
+            .send(FileSystemCommand(FileSystemCommandInner::DirIdbDrop(
+                idb_key, oneshot_tx,
+            )))
+            .unwrap();
+        oneshot_rx.recv().unwrap()
+    }
+
+    /// Returns a path consisting of a single element: the name of the root directory of this
+    /// filesystem.
+    pub fn root_path(&self) -> &camino::Utf8Path {
+        self.name.as_str().into()
+    }
+
+    /// Returns the key needed to restore this `FileSystem` using `FileSystem::from_idb()`.
+    pub fn idb_key(&self) -> Option<&str> {
+        self.idb_key.as_deref()
+    }
+}
+
+impl Drop for FileSystem {
+    fn drop(&mut self) {
+        let (oneshot_tx, oneshot_rx) = oneshot::channel();
+        filesystem_tx_or_die()
+            .send(FileSystemCommand(FileSystemCommandInner::DirDrop(
+                self.key, oneshot_tx,
+            )))
+            .unwrap();
+        oneshot_rx.recv().unwrap();
+    }
+}
+
+impl Clone for FileSystem {
+    fn clone(&self) -> Self {
+        let (oneshot_tx, oneshot_rx) = oneshot::channel();
+        filesystem_tx_or_die()
+            .send(FileSystemCommand(FileSystemCommandInner::DirClone(
+                self.key, oneshot_tx,
+            )))
+            .unwrap();
+        Self {
+            key: oneshot_rx.recv().unwrap(),
+            name: self.name.clone(),
+            idb_key: self.idb_key.clone(),
+        }
+    }
+}
+
+impl FileSystemTrait for FileSystem {
+    type File = File;
+
+    fn open_file(
+        &self,
+        path: impl AsRef<camino::Utf8Path>,
+        mut flags: OpenFlags,
+    ) -> Result<Self::File> {
+        if flags.contains(OpenFlags::Truncate) || flags.contains(OpenFlags::Create) {
+            flags |= OpenFlags::Write;
+        }
+        let (oneshot_tx, oneshot_rx) = oneshot::channel();
+        filesystem_tx_or_die()
+            .send(FileSystemCommand(FileSystemCommandInner::DirOpenFile(
+                self.key,
+                path.as_ref().to_path_buf(),
+                flags,
+                oneshot_tx,
+            )))
+            .unwrap();
+        oneshot_rx.recv().unwrap().map(|key| File { key })
+    }
+
+    fn metadata(&self, path: impl AsRef<camino::Utf8Path>) -> Result<Metadata> {
+        let (oneshot_tx, oneshot_rx) = oneshot::channel();
+        filesystem_tx_or_die()
+            .send(FileSystemCommand(FileSystemCommandInner::DirEntryMetadata(
+                self.key,
+                path.as_ref().to_path_buf(),
+                oneshot_tx,
+            )))
+            .unwrap();
+        oneshot_rx.recv().unwrap()
+    }
+
+    fn rename(
+        &self,
+        from: impl AsRef<camino::Utf8Path>,
+        to: impl AsRef<camino::Utf8Path>,
+    ) -> Result<()> {
+        Err(Error::NotSupported)
+    }
+
+    fn exists(&self, path: impl AsRef<camino::Utf8Path>) -> Result<bool> {
+        let (oneshot_tx, oneshot_rx) = oneshot::channel();
+        filesystem_tx_or_die()
+            .send(FileSystemCommand(FileSystemCommandInner::DirEntryExists(
+                self.key,
+                path.as_ref().to_path_buf(),
+                oneshot_tx,
+            )))
+            .unwrap();
+        Ok(oneshot_rx.recv().unwrap())
+    }
+
+    fn create_dir(&self, path: impl AsRef<camino::Utf8Path>) -> Result<()> {
+        let (oneshot_tx, oneshot_rx) = oneshot::channel();
+        filesystem_tx_or_die()
+            .send(FileSystemCommand(FileSystemCommandInner::DirCreateDir(
+                self.key,
+                path.as_ref().to_path_buf(),
+                oneshot_tx,
+            )))
+            .unwrap();
+        oneshot_rx.recv().unwrap()
+    }
+
+    fn remove_dir(&self, path: impl AsRef<camino::Utf8Path>) -> Result<()> {
+        let (oneshot_tx, oneshot_rx) = oneshot::channel();
+        filesystem_tx_or_die()
+            .send(FileSystemCommand(FileSystemCommandInner::DirRemoveDir(
+                self.key,
+                path.as_ref().to_path_buf(),
+                oneshot_tx,
+            )))
+            .unwrap();
+        oneshot_rx.recv().unwrap()
+    }
+
+    fn remove_file(&self, path: impl AsRef<camino::Utf8Path>) -> Result<()> {
+        let (oneshot_tx, oneshot_rx) = oneshot::channel();
+        filesystem_tx_or_die()
+            .send(FileSystemCommand(FileSystemCommandInner::DirRemoveFile(
+                self.key,
+                path.as_ref().to_path_buf(),
+                oneshot_tx,
+            )))
+            .unwrap();
+        oneshot_rx.recv().unwrap()
+    }
+
+    fn read_dir(&self, path: impl AsRef<camino::Utf8Path>) -> Result<Vec<DirEntry>> {
+        let (oneshot_tx, oneshot_rx) = oneshot::channel();
+        filesystem_tx_or_die()
+            .send(FileSystemCommand(FileSystemCommandInner::DirReadDir(
+                self.key,
+                path.as_ref().to_path_buf(),
+                oneshot_tx,
+            )))
+            .unwrap();
+        oneshot_rx.recv().unwrap()
+    }
+}
+
+impl Drop for File {
+    fn drop(&mut self) {
+        let (oneshot_tx, oneshot_rx) = oneshot::channel();
+        filesystem_tx_or_die()
+            .send(FileSystemCommand(FileSystemCommandInner::FileDrop(
+                self.key, oneshot_tx,
+            )))
+            .unwrap();
+        oneshot_rx.recv().unwrap();
+    }
+}
+
+impl crate::File for File {
+    fn metadata(&self) -> crate::Result<Metadata> {
+        let (oneshot_tx, oneshot_rx) = oneshot::channel();
+        filesystem_tx_or_die()
+            .send(FileSystemCommand(FileSystemCommandInner::FileSize(
+                self.key, oneshot_tx,
+            )))
+            .unwrap();
+        let size = oneshot_rx.recv().unwrap();
+        Ok(Metadata {
+            is_file: true,
+            size,
+        })
+    }
+}
+
+impl std::io::Read for File {
+    fn read(&mut self, buf: &mut [u8]) -> std::io::Result<usize> {
+        let (oneshot_tx, oneshot_rx) = oneshot::channel();
+        filesystem_tx_or_die()
+            .send(FileSystemCommand(FileSystemCommandInner::FileRead(
+                self.key,
+                buf.len(),
+                oneshot_tx,
+            )))
+            .unwrap();
+        let vec = oneshot_rx.recv().unwrap()?;
+        let length = vec.len();
+        buf[..length].copy_from_slice(&vec[..]);
+        Ok(length)
+    }
+}
+
+impl std::io::Write for File {
+    fn write(&mut self, buf: &[u8]) -> std::io::Result<usize> {
+        let (oneshot_tx, oneshot_rx) = oneshot::channel();
+        filesystem_tx_or_die()
+            .send(FileSystemCommand(FileSystemCommandInner::FileWrite(
+                self.key,
+                buf.to_vec(),
+                oneshot_tx,
+            )))
+            .unwrap();
+        oneshot_rx.recv().unwrap()?;
+        Ok(buf.len())
+    }
+
+    fn flush(&mut self) -> std::io::Result<()> {
+        let (oneshot_tx, oneshot_rx) = oneshot::channel();
+        filesystem_tx_or_die()
+            .send(FileSystemCommand(FileSystemCommandInner::FileFlush(
+                self.key, oneshot_tx,
+            )))
+            .unwrap();
+        oneshot_rx.recv().unwrap()
+    }
+}
+
+impl std::io::Seek for File {
+    fn seek(&mut self, pos: std::io::SeekFrom) -> std::io::Result<u64> {
+        let (oneshot_tx, oneshot_rx) = oneshot::channel();
+        filesystem_tx_or_die()
+            .send(FileSystemCommand(FileSystemCommandInner::FileSeek(
+                self.key, pos, oneshot_tx,
+            )))
+            .unwrap();
+        oneshot_rx.recv().unwrap()
+    }
+}
